(search) move normalizeAggregations.js to lib and update references
(cherry picked from commit 86380fd)
