+++ applications/search/src/api/get-datasets.js
@@ -2,8 +2,7 @@ import _ from 'lodash';
 import axios from 'axios';
 
 import { addOrReplaceParam } from '../lib/addOrReplaceUrlParam';
-import { normalizeAggregations } from './normalizeAggregations';
-// import datasetsApiResponseFixture from '../pages/search-page/__fixtures/datasetsApiResponse.json'
+import { normalizeAggregations } from './../lib/normalizeAggregations';
 
 export const getDatasets = async search => {
   const datasetsUrl = `/datasets${search}`;
@@ -13,7 +12,6 @@ export const getDatasets = async search => {
     .get(url)
     .catch(e => console.log(JSON.stringify(e))); // eslint-disable-line no-console
 
-  // return normalizeAggregations(datasetsApiResponseFixture);
   return response && normalizeAggregations(response.data);
 };
