+++ applications/search/src/pages/search-page/results-dataset/search-hit-item/search-hit-item.component.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { SearchHitItem } from './search-hit-item.component';
 import datasetsResponse from '../../__fixtures/datasetsApiResponse.json';
-import { normalizeAggregations } from '../../../../api/normalizeAggregations';
+import { normalizeAggregations } from '../../../../lib/normalizeAggregations';
 
 const datasetItems = normalizeAggregations(datasetsResponse);
